rename: following, like, support, host폴더 생성
