+++ src/src/components/my-fund/like/MyHeartList.jsx
@@ -1,6 +1,6 @@
 import { Suspense } from "react";
 import InfiniteFundInfo from "@/components/fund-list/InfiniteFundInfo";
-import InfiniteFundInfoLoader from "../fund-list/InfiniteFundInfo.loader";
+import InfiniteFundInfoLoader from "../../fund-list/InfiniteFundInfo.loader";
 
 function MyHeartList() {
   return (
